updater: fix enum value `kInstaller` to be `7` to match `mojom`
This CL fixes an issue of the `mojom` having an additional value of
`kUnknown` with an `enum` value of `6`, thereby making the value of
`kInstaller` to be `7` in the `mojom`, but mismatched with the rest of
the codebase that has the value of `kInstaller` at `6`.

This is a fix for r1315365, which introduced a new error category of
`kInstaller` to separate the `update_client::InstallError` errors from
errors returned by the application installers.

Bug: b/40259471
Change-Id: I9ea989a0c8a5bed644305bcf476b9b2f24390e23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5634373
Commit-Queue: S Ganesh <ganesh@chromium.org>
Auto-Submit: S Ganesh <ganesh@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315429}