[Tab Group Sync] Update creator_cache_guid for tabs
This CL updates the creator_cache_guid for tabs at the same time when
the creator_cache_guid for groups is updated. Earlier CLs have
invoked the UpdateLocalCacheGuid method on the model whenever a tab
or a group is created. This CL simply adds the implementation for tabs.
Also refactors out the loop in the bridge to its helper function.

Bug: 344873349
Change-Id: I2395f3f812e2e2aeebea9f51178ffdc6a39ad803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5634292
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315433}