[Downloads] Fix referrer line displaying on downloads without a display referrer URL.
Downloads without a referrer URL will have a display referrer URL that
is an empty string. This is causing the referrer line to still display
which is incorrect. The intended behavior is for no referrer line to
show on downloads without a referrer URL.

The else clause is necessary because the <iron-list> wrapper in
manager.ts causes a download without a referrer URL to use the previous
download's URL (if it exists).

Bug: 346472690
Change-Id: I5f3564c54e689ebb2d9238100aed2ae1df5b4caf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5634099
Reviewed-by: Lily Chen <chlily@chromium.org>
Commit-Queue: Kerry Meng-Lin <kmenglin@google.com>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315439}