Use GetOrCreateForDocument in ApplyFeatureDiffForOriginTrial
Currently in OriginTrialStateHostImpl::ApplyFeatureDiffForOriginTrial() there is a CHECK(document_data) that occasionally fails. When the CHECK fails, it indicates that RuntimeFeatureStateDocumentData:: GetForCurrentDocument() is returning a nullptr. This should not be happening, as all RenderFrameHosts should have RuntimeFeatureStateDocumentData created for it prior to commit.

To avoid failing this CHECK frequently, we can use RuntimeFeatureStateDocumentData::GetOrCreateForCurrentDocument() instead, which will avoid us encountering a nullptr dereference.

A follow-up bug has been filed to determine the root cause: crbug.com/347186599.

Bug: 344509105
Change-Id: I6c2080af077a598da31b48854d65cfcf3c04201d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5631365
Reviewed-by: Steven Bingler <bingler@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315443}