[Tab Group Sync] Provide kill switch for auto open
This CL adds a kill switch for auto-open which can be used in
case of emergency to disable auto-open tab groups. The user
settings for auto-open will still be visible, but the user
selection will be ignored. This significantly reduces the surface
area for risky functionalities.

Bug: 347132506
Change-Id: I2fa074a65f120e620ac58d72eb2a739181843f2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5633200
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315444}