[Hub] Remove the AndroidHub flag
This CL is the first step in removing the AndroidHub flag. It removes
the feature flag entry. As a consequence a number of tests now become
invalid or reference unreachable code and are deleted.

In a follow up CL HubFieldTrial#isHubEnabled() calls will be removed.

Bug: 345524466
Change-Id: I11cf287462274ef43a6db15379205537e7c777c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5629936
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315470}