[views-ax] Use ViewAccessibility::SetName in ui/*
This CL migrates callers in ui/* from using View::SetAccessibleName to
the new ViewAccessibility::SetName. View::SetAccessibleName has 500+
call sites. In order to make changes to all of these call sites easier
to review, updates will be broken down into multiple CLs based on the
filepath of the changes.

This change does not introduce any behavior change and is just a
rename.

This is the third CL to use ViewAccessibility::SetName. Follow-up
changes will migrate the other callers of View::SetAccessibleName.

The previous CLs migrated the calls in other directories:
CL:5579219
CL:5594199

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

AX-Relnotes: n/a

Bug: 325137417
Change-Id: I3739070212164f270300a676f5f2a6a78b3daf02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5619292
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Evelynn Kaplan <evelynn.kaplan@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1315486}