Remove same-size requirement from HardwareDisplayController seamless mode test
This change removes a requirement for seamless mode testing that the
mode-under-test must have the same size as the current mode. That case
is supported by some hardware, so the check is not needed. Other
hardware should not be affected, as the test modesets will still fail
in those cases. Plus the check persists at higher levels of abstraction
where it is more appropriate.

This change also renames the TestSeamlessRefreshRate method to
TestSeamlessMode, which more accurately reflects its intended usage.

Bug: 321725211
Change-Id: I76f4886535ad3bbf11050a22480903f0c44b4875
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5622191
Commit-Queue: Andrew Wolfers <aswolfers@chromium.org>
Reviewed-by: Gil Dekel <gildekel@chromium.org>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315600}