[Autofill] CHECK that fields are part of forms in AutofillManager events
Suppose in a renderer -> browser event (like AskForValuesToFill()) the
FieldGlobalId does not identify a FormFieldData of the FormData.

Before this CL, we early returned in [Browser]AutofillManager in this
case.

After this CL, we CHECK that this scenario does not happen. This
property is now guaranteed by ContentAutofillDriver and by
AutofillDriverIOS (and its AutofillAgent).

Bug: 40232021
Change-Id: I7409d07b1031aea76bf301e2c4964d4f99734f22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5630374
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1315610}