Move some HttpBasicStream methods to HttpBasicState
This CL moves implementations of the following methods from
HttpBasicStream and WebSocketHandshakeStream to HttpBasicState:
* Close()
* {Set,Can}ConnectionReused()
* GetLoadTimingInfo()
* GetSSLInfo()
* GetRemoteEndpoint()

This refactoring is to remove ClientSocketHandle dependencies from
HttpBasicStream so that we can use HttpBasicStream without
ClientSocketHandle. Note that HttpBasicStream still depends on
ClientSocketHandle. I'll try to remove the dependencies in
later CLs.

Bug: 346835898
Change-Id: Iea4c7ee14d66e6e8ab5cbfb69bdfa739363ca4a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5632915
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315620}