[Tab Group Sync] Remove redundant messages to sync on local ID change
Local IDs are a local only concept and stored in sync local DB.
They shouldn't trigger an update in the update timestamp or should
they trigger a message to sync. This CL rectifies this behavior
which was broken and always sending a message to sync when a group
is opened or closed. We still have a message to sync on tab group
open in desktop due to navigation observer. But that's a bit harder
to fix. Moreover, this behavior was wrongly introduced recently in
This behavior was added by
a recent CL
https://chromium-review.googlesource.com/c/chromium/src/+/5583203/14/components/saved_tab_groups/saved_tab_group_sync_bridge.cc
This CL fixes that.

Bug: 347585919
Change-Id: Ifa613aef0a40c39c36cba30b52d3d6a184ee87d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5634484
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315623}