Remove a duplicated function `FirstAvailableFont()`
crrev.com/c/5579080 added a function for crrev.com/c/5577054
to find the first available font. It turned out that the
existing function `FontList::FirstAvailableOrFirst()` does the
same work, used for the same purpose in a different code path.

There is a subtle difference for an empty input, but this code
path uses the function only for strings in the resource, and
that the difference doesn't matter.

This patch changes to use the existing function and removes
the duplicated function, unifying the macOS code path added in
crrev.com/c/5633154.

This patch has no behavior changes.

Bug: 41323186
Change-Id: Ic17a70e75ed5f4ffe069564bb545e60c972369d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5632837
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315644}