Always show Picker back button when a category is selected.
Currently, the back button visibility is updated when the active Picker
page changes (it's shown when the category page is the active page).
However, this doesn't exactly correspond to when the user has selected a
category, since when the user starts searching, the active page switches
to the "search results page", regardless of whether a category is
selected.

To fix this, show the back button when the category is initially
selected, and hide it only after the user clicks it to go back to the
zero state page.

Bug: b:347284310
Change-Id: If3744f32456b73b8126bc60fb5bb80e14ddb8c6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5632090
Commit-Queue: Michelle Chen <michellegc@google.com>
Reviewed-by: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315671}