Don't annotate list role to collection info without count
CollectionInfo might be missing count information. ChromeVox doesn't
expect a list without count information. We don't want to announce it as
a list in that case.

information.

Bug: b:323423298
Test: AccessibilityNodeInfoDaraWrapperTests
AX-Relnotes: ChromeVox won't announce the size of list on a list which doesn't have count in Android
Change-Id: I02ae2b12636f8261d00ba463c535aad088591386
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5602551
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Commit-Queue: Saho Kobayashi <sahok@chromium.org>
Reviewed-by: Jacob Francis <francisjp@google.com>
Cr-Commit-Position: refs/heads/main@{#1315713}