[Uno-d] Update IdentityGetAuthTokenFunction to use Signin PA
This CL simplifies `IdentityGetAuthTokenFunction` to prepare for
supporting signed in on the web only users. It shifts to using the Sign
in primary account, this is required to pre-fill the reauth Gaia sign in
page with the correct email address in case of sign in paused state.

Bug: 347169207
Change-Id: Ie34d701cb4a6dca3aa31b76d1d05b60bb9fa4b15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5630656
Reviewed-by: David Roger <droger@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315731}