[iOS][QD] Use QuickDeleteCommands on all CBD callers
Most of clear browsing data callers rely on the SettingsCommands to open
CBD. In these cases, we have simply switched to rely on
QuickDeleteCommands directly. However, the history page wraps the CBD
UI in a controller instead of relying on the SettingsCommands. We don't
see a reason to continue to follow this pattern: the privacy settings
also shows CBD or QD on top of the another view. As such, this CL also
makes the history page rely directly on the QuickDeleteCommands.

This CL also disables the keyboard shortcut to open Clear Browsing Data
or Quick Delete from an incognito window (approved by incognito's PM,
andrewkamau@).

Fixed: 340796825
Bug: 335387869
Change-Id: Iad41ca327e8360770362b235a3faf466437a4c81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5626059
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Filipa Senra <fsenra@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1315733}