[UB] Update UB toggle to be ON when protections are ON.
This is the opposite of the current UI. Now, when a user turns off
the toggle, a new site exception is created. experiment gated with
`kTrackingProtectionSettingsLaunch` flag.

Also updated the `allow_third_party_cookies` param to be more specific
to toggle logic: `is_toggled`.

Bug: b:344856056
Change-Id: Ic1f8a4f6031f43ce74c0d21f73ede3023fbe9b9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5630894
Commit-Queue: Michelle Abreo <michelleabreo@google.com>
Reviewed-by: Fiona Macintosh <fmacintosh@google.com>
Reviewed-by: Andrey Zaytsev <andzaytsev@google.com>
Cr-Commit-Position: refs/heads/main@{#1315785}