[iOS] Download pdf improvements -- followup
The aim of this CL is to handle the files omnibox sharing. The files
sharing is enabled when the file's URL follows either
`chrome://external-file` or `chrome://downloads`.
In addition this CL removes the `Save` bar for local storage files.
And when a download is triggered from a web page the `OPEN IN...` is not
replaced by the `OPEN` button.

Bug: 344849240, 344849291, 344849291
Change-Id: I493c49c5edc0c432c30fda3e629e92a73049725c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5614030
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Elmehdi Rahmaoui <erahmaoui@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1315834}