[Uno-d] Show Chrome Sign in dialog for extensions
This CL shows the Chrome Sign in dialog for extensions. This is
triggered if the extension starts the interactive Sign in flow through
the identity API while the user is only signed in on the Web.

Bug: 347296107
Change-Id: I35739543d99cdff8735745322ca78b9a5dd16568
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5633481
Reviewed-by: David Roger <droger@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315839}