[iOS][CBD] Invalid dependencies of browser when browser is destroyed
We're seeing crashes of invalid address on clear browsing data. We
believe this is because either the controller is being stopped after the
browser state has been destroyed or that is not stopped when the browser
state is destroyed. To avoid dangling pointers in all classes
initialised in the CBD UI with browser (or derived objects), we have
decided to observe the Browser and shut everything when it's about to be
destroyed. As an extra layer of protection, we've also switched the
pointers associated with Browser/BrowserState to weak pointers.

Fixed: 344602027, 345692813
Change-Id: I323f9c048fd0e2df5258d3b79b84fc621107991d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5612399
Commit-Queue: Filipa Senra <fsenra@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1315845}