[ios] Temporary fix test WebStateTest.RestorePageTitles
This is a temporary solution. Tests are failing because SetTitle() uses
a property deprecated in iOS18.

Proposed fix for this problem is to use GetTitleForDisplay() instead
which gets the title from the URL if empty.
For additional details check the bug description.

Bug: 347207400
Change-Id: I1aea9839f7f442091461496f405cc7b489fa22d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5630686
Commit-Queue: Federica Germinario <fedegermi@google.com>
Auto-Submit: Federica Germinario <fedegermi@google.com>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315867}