Fix unsafe-buffers warnings in checked iterators
The iterator is an abstraction that does bounds checking over pointers
so it must use pointers and UNSAFE_BUFFERS() internally. Write safety
comments explaining the soundness of its operations.

Its constructors also take pointers so require unsafe pointer usage to
call. Write some helpers in the unit tests to isolate the pointering
and that CHECK for validity.

Add death tests for going out of bounds with the iterator.

Change-Id: I49fe5860a349fdc97e9a1acf1b34dbd193a1ccd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5634152
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1315876}